pyproject.toml
[project]
name = "ghai"
version = "0.1.0"
description = "AI-powered Git/GitHub tools"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "keyring>=25.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
ghai = "ghai.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ghai"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ghai"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
explicit_package_bases = true
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = false
exclude = [
    "^\\.venv",
    "^venv",
    "^build",
    "^dist",
]
